🐛 fix(rebase): compatibility fallout
